// libs/rc/shared/utils/src/models/OsVendorEnum.ts
export enum OsVendorEnum {
  All = 'All',
  Windows = 'Windows',
  MacOs = 'macOS',
  ChromeOs = 'ChromeOs',
  Linux = 'Linux',
  Ubuntu = 'Ubuntu',
  Ios = 'iOS',
  Android = 'Android',
  BlackBerry = 'BlackBerry',
  AmazonKindle = 'Amazon Kindle',
  CiscoIpPhone = 'Cisco',
  AvayaIpPhone = 'Avaya',
  LinksysPapVoip = 'LinksysPapVoip',
  NortelIpPhone = 'Nortel',
  /**
   * @deprecated
   */
  Xbox360 = 'Xbox 360',
  /**
   * @deprecated
   */
  PlayStation2 = 'PlayStation2',
  GameCube = 'GameCube',
  Wii = 'Wii',
  /**
   * @deprecated
   */
  PlayStation3 = 'PlayStation3',
  PlayStation = 'PlayStation',
  Xbox = 'Xbox',
  Nintendo = 'Nintendo',
  HpPrinter = 'HP',
  CanonPrinter = 'Canon',
  XeroxPrinter = 'Xerox',
  DellPrinter = 'Dell',
  BrotherPrinter = 'Brother',
  EpsonPrinter = 'Epson',
  NestCamera = 'Nest Camera',
  NestThermostat = 'Nest Thermostat',
  WemoSmartSwitch = 'Wemo Smart Switch',
  WifiSmartPlug = 'Wi-Fi Smart Plug',
  SonyPlayer = 'SonyPlayer',
  PanasonicG20Tv = 'PanasonicG20Tv',
  SamsungSmartTv = 'SamsungSmartTv',
  AppleTv = 'AppleTv',
  LibratoneSpeakers = 'LibratoneSpeakers',
  BoseSpeakers = 'BoseSpeakers',
  SonosSpeakers = 'SonosSpeakers',
  RokuStreamingStick = 'RokuStreamingStick',
  TelenetCpe = 'TelenetCpe',
}